Update separating navbar
